move Scheds into a package
This way Sched can inherits from SchedV2 as in anki, using protected
variables and method betwen them
