+++ AnkiDroid/src/main/java/com/ichi2/anki/widgets/DeckAdapter.java
@@ -34,7 +34,7 @@ import android.widget.TextView;
 import com.ichi2.anki.R;
 import com.ichi2.compat.CompatHelper;
 import com.ichi2.libanki.Collection;
-import com.ichi2.libanki.Sched;
+import com.ichi2.libanki.sched.Sched;
 
 import com.ichi2.utils.JSONObject;
