+++ AnkiDroid/src/main/java/com/ichi2/libanki/Collection.java
@@ -34,6 +34,9 @@ import com.ichi2.async.DeckTask;
 import com.ichi2.compat.CompatHelper;
 import com.ichi2.libanki.exception.NoSuchDeckException;
 import com.ichi2.libanki.hooks.Hooks;
+import com.ichi2.libanki.sched.AbstractSched;
+import com.ichi2.libanki.sched.Sched;
+import com.ichi2.libanki.sched.SchedV2;
 import com.ichi2.libanki.template.Template;
 import com.ichi2.upgrade.Upgrade;
 import com.ichi2.utils.FunctionalInterfaces;
